Rename the iOS project from FileActions to Fileaway (#77)
This change updates the name of the iOS Xcode project and core libraries to 'Fileaway'. It had previously been renamed to FileActions / File Actions in an attempt to find a name that wasn't already taken on the App Store, but that's a terrible reason to rename a project. With this, the macOS and iOS projects should have the same name, leaving just the core library to be renamed.

It also introduces a drive-by update to the README and an iOS screenshot, acknowledging the significant differences in functionality between the iOS and macOS apps.