Signed and notarized automated macOS builds (#89)
This change goes some way to automating the full release process by updating the build script to automatically sign and notarise the macOS builds. While it makes use of Fastlane's `notarize` and `match` actions, it doesn't migrate the whole build process to Fastlane (I tried that) because it just seems to add unnecessary complexity as Fastlane's trying to do too much magic under the hood and guesses wrong.

Longer-term I'd like to add support for using some flavour of semantic versioning and conventional commits to auto-generate the release notes and version number.