feat(macOS): Multiple inbox locations (#129)
This changes adds support for multiple inbox and archive locations. In the process, it significantly reworks the way directory observers are managed in the Manager to ensure the app dynamically updates when the locations change.

This change also includes a drive-by fix that ensures the Rules Wizard is accessing details from the RuleSet, and not the Settings instance directly.