fix(macOS): Duplicate and delete context menu items act on all selected rules in the Rules editor (#140)
This change introduces a new SwiftUI View, `MenuItem` that correctly determines the items that the current context menu should act upon. This is necessary as the loop item in the `ForEach` isn't actually the focused item--macOS uses the loop item _if_ there is no existing selection or the loop item is outside the current selection, or the selected items if the loop item / right click item is within the selection.