fix(macOS): Set the initial directory when picking rules destinations
The directory picker in the Rules editor didn’t have a default location, meaning it opened in the last directory the user had used. This is confusing and makes it really easy to mistakenly pick a directory that’s not contained within the archive root. This change uses the existing directory if it exists, and the root of the archive otherwise.