feat: Sort by the file creation date if no date can be found in the filename
This change also moves caching into the DirectoryObserver and out of the DateFinder and TitleFinder utilities to better match the lifecycle of the data and sure up thread safety.

In order to better help users understand where the dates are coming from, it adds a tooltip to the date element in the list.