fix: Directory views don't notice changes when moving or deleting folders (#517)
Ensure `DirectoryMonitor` correctly indexes newly-added folders, and
deletes files contained within removed folders. As part of this fix, the
change introduces unit tests for `DirectoryMonitor` and
`DirectoryViewModel`.

This includes a drive-by fix for selection management when viewing
documents using QuickLook: unfortunately it seems like there's a bug
with the SwiftUI QuickLook bindings on iOS that means the current item
binding isn't updated when paging through items in the QuickLook view.
Given this, it's not possible to ensure the selection is correct, so we
simply allow the selection to be cleared when the item binding goes nil
(indicating that the QuickLook view has been dismissed).